ci: add dependabot dependency updates (#9)
This pull request updates the `.github/dependabot.yml` file to configure
Dependabot for automatic dependency updates. The most important change
is the addition of a new configuration for npm package updates.

Configuration for automatic dependency updates:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R1-R6):
Added configuration for Dependabot to check for npm package updates on a
weekly schedule.