ci: add eslint github workflow (#15)
This pull request introduces a new GitHub Actions workflow to run ESLint
on the codebase. The workflow is triggered on pushes and pull requests
to the `main` branch, as well as on a scheduled basis. The most
important changes include the addition of the workflow configuration and
the steps to install dependencies, run ESLint, and upload the results.

New GitHub Actions workflow:

*
[`.github/workflows/eslint.yml`](diffhunk://#diff-8deb4e00ff9e68f9e5e25914b8d329a413bfbc90ab5f8c97662324cc68903d2fR1-R51):
Added a new workflow named `ESLint` that runs on pushes and pull
requests to the `main` branch, and on a weekly schedule. The workflow
installs dependencies, runs ESLint, and uploads the analysis results to
GitHub.